Crate iota_client

A general purpose IOTA client library for interaction with the IOTA network (Tangle)

High-level functions are accessible via the Client.

Sending a message with an indexation payload

let iota = Client::builder()
  .with_node("https://api.lb-0.testnet.chrysalis2.com")?
  .finish()
  .await?;

let message = iota
  .message()
  .with_index("Hello")
  .with_data("Tangle".as_bytes().to_vec())
  .finish()
  .await?;

println!("Message sent {}", message.id().0);

Re-exports

pub use bee_common as common;

pub use bee_message;

pub use bee_pow as pow;

pub use bee_rest_api;

pub use builder::ClientBuilder;

pub use client::*;

pub use crypto;

pub use error::*;

pub use storage::*;

Modules

api
High level APIs

async_runtime

builder
Builder of the Client Instance

client
The Client module to connect through HORNET or Bee with API usages

error
Error handling in iota-client crate.

node
Iota node APIs

node_manager
The node manager that takes care of sending requests and quroum if enabled

storage
Storage adapter

Structs

AddressOutputsOptions
The outputs query options.

Seed
A seed is an arbitrary bytestring used to create the root of the tree.

Topic
A MQTT topic.

Url
A parsed URL record.

Enums

OutputType
Output type filter.