Module iota_client::api::consolidation

Functions

consolidate_funds
Function to consolidate all funds from a range of addresses to the address with the lowest index in that range Returns the address to which the funds got consolidated, if any were available