Struct iota_client::client::Client

pub struct Client {
    pub(crate) runtime: Option<Runtime>,
    pub(crate) node_manager: NodeManager,
    pub(crate) sync_kill_sender: Option<Arc<Sender<()>>>,
    pub(crate) mqtt_client: Option<MqttClient>,
    pub(crate) mqtt_topic_handlers: Arc<RwLock<HashMap<Topic, Vec<Arc<Box<dyn Fn(&TopicEvent) + Send + Sync>>>>>>,
    pub(crate) broker_options: BrokerOptions,
    pub(crate) mqtt_event_channel: (Arc<WatchSender<MqttEvent>>, WatchReceiver<MqttEvent>),
    pub(crate) network_info: Arc<RwLock<NetworkInfo>>,
    pub(crate) request_timeout: Duration,
    pub(crate) api_timeout: HashMap<Api, Duration>,
}

An instance of the client using HORNET or Bee URI

Fields
runtime: Option<Runtime>node_manager: NodeManager

Node manager

sync_kill_sender: Option<Arc<Sender<()>>>

Flag to stop the node syncing

mqtt_client: Option<MqttClient>

A MQTT client to subscribe/unsubscribe to topics.

mqtt_topic_handlers: Arc<RwLock<HashMap<Topic, Vec<Arc<Box<dyn Fn(&TopicEvent) + Send + Sync>>>>>>broker_options: BrokerOptionsmqtt_event_channel: (Arc<WatchSender<MqttEvent>>, WatchReceiver<MqttEvent>)network_info: Arc<RwLock<NetworkInfo>>request_timeout: Duration

HTTP request timeout.

api_timeout: HashMap<Api, Duration>

HTTP request timeout for each API call.

Implementations

impl Client

pub fn builder() -> ClientBuilder

Create the builder to instntiate the IOTA Client.

pub(crate) fn start_sync_process(
runtime: &Runtime,
sync: Arc<RwLock<HashSet<Url>>>,
nodes: HashSet<Url>,
node_sync_interval: Duration,
network_info: Arc<RwLock<NetworkInfo>>,
kill: Receiver<()>
)

Sync the node lists per node_sync_interval milliseconds

pub(crate) async fn sync_nodes(
sync: &Arc<RwLock<HashSet<Url>>>,
nodes: &HashSet<Url>,
network_info: &Arc<RwLock<NetworkInfo>>
)

pub(crate) async fn get_node(&self) -> Result<Url>

Get a node candidate from the synced node pool.

pub async fn get_network_id(&self) -> Result<u64>

Gets the network id of the node we're connecting to.

pub async fn get_pow_provider(&self) -> ClientMiner

Gets the miner to use based on the PoW setting

pub async fn get_network_info(&self) -> Result<NetworkInfo>

Gets the network related information such as network_id and min_pow_score and if it's the default one, sync it first.

pub async fn get_bech32_hrp(&self) -> Result<String>

returns the bech32_hrp

pub async fn get_min_pow_score(&self) -> Result<f64>

returns the min pow score

pub async fn get_tips_interval(&self) -> u64

returns the tips interval

pub async fn get_local_pow(&self) -> bool

returns the local pow

pub async fn unsynced_nodes(&self) -> HashSet<&Url>

returns the unsynced nodes.

pub fn subscriber(&mut self) -> MqttManager<'_>

Returns a handle to the MQTT topics manager.

pub fn mqtt_event_receiver(&self) -> WatchReceiver<MqttEvent>

Returns the mqtt event receiver.

pub(crate) fn get_timeout(&self, api: Api) -> Duration

pub async fn get_node_health(url: &str) -> Result<bool>

GET /health endpoint

pub async fn get_health(&self) -> Result<bool>

GET /health endpoint

pub async fn get_node_info(
url: &str,
auth_name_passw: Option<(&str, &str)>
) -> Result<NodeInfo>

GET /api/v1/info endpoint

pub async fn get_info(&self) -> Result<NodeInfoWrapper>

GET /api/v1/info endpoint

pub async fn get_peers(&self) -> Result<Vec<PeerDto>>

GET /api/v1/peers endpoint

pub async fn get_tips(&self) -> Result<Vec<MessageId>>

GET /api/v1/tips endpoint

pub async fn post_message(&self, message: &Message) -> Result<MessageId>

POST /api/v1/messages endpoint

pub async fn post_message_json(&self, message: &Message) -> Result<MessageId>

POST JSON to /api/v1/messages endpoint

pub fn get_message(&self) -> GetMessageBuilder<'_>

GET /api/v1/messages/{messageId} endpoint

pub async fn get_output(&self, output_id: &UtxoInput) -> Result<OutputResponse>

GET /api/v1/outputs/{outputId} endpoint Find an output by its transaction_id and corresponding output_index.

pub async fn find_outputs(
&self,
outputs: &[UtxoInput],
addresses: &[String]
) -> Result<Vec<OutputResponse>>

Find all outputs based on the requests criteria. This method will try to query multiple nodes if the request amount exceeds individual node limit.

pub fn get_address(&self) -> GetAddressBuilder<'_>

GET /api/v1/addresses/{address} endpoint

pub async fn get_milestone(&self, index: u32) -> Result<MilestoneResponse>

GET /api/v1/milestones/{index} endpoint Get the milestone by the given index.

pub async fn get_milestone_utxo_changes(
&self,
index: u32
) -> Result<MilestoneUTXOChanges>

GET /api/v1/milestones/{index}/utxo-changes endpoint Get the milestone by the given index.

pub async fn get_receipts(&self) -> Result<Vec<ReceiptDto>>

GET /api/v1/receipts endpoint Get all receipts.

pub async fn get_receipts_migrated_at(
&self,
milestone_index: u32
) -> Result<Vec<ReceiptDto>>

GET /api/v1/receipts/{migratedAt} endpoint Get the receipts by the given milestone index.

pub async fn get_treasury(&self) -> Result<TreasuryResponse>

GET /api/v1/treasury endpoint Get the treasury output.

pub async fn get_included_message(
&self,
transaction_id: &TransactionId
) -> Result<Message>

GET /api/v1/transactions/{transactionId}/included-message Returns the included message of the transaction.

pub async fn reattach(
&self,
message_id: &MessageId
) -> Result<(MessageId, Message)>

Reattaches messages for provided message id. Messages can be reattached only if they are valid and haven't been confirmed for a while.

pub async fn reattach_unchecked(
&self,
message_id: &MessageId
) -> Result<(MessageId, Message)>

Reattach a message without checking if it should be reattached

pub async fn promote(
&self,
message_id: &MessageId
) -> Result<(MessageId, Message)>

Promotes a message. The method should validate if a promotion is necessary through get_message. If not, the method should error out and should not allow unnecessary promotions.

pub async fn promote_unchecked(
&self,
message_id: &MessageId
) -> Result<(MessageId, Message)>

Promote a message without checking if it should be promoted

pub fn message(&self) -> ClientMessageBuilder<'_>

A generic send function for easily sending transaction or indexation messages.

pub fn get_unspent_address<'a>(
&'a self,
seed: &'a Seed
) -> GetUnspentAddressBuilder<'a>

Return a valid unspent address.

pub fn get_addresses<'a>(&'a self, seed: &'a Seed) -> GetAddressesBuilder<'a>

Return a list of addresses from the seed regardless of their validity.

pub async fn find_messages<I: AsRef<[u8]>>(
&self,
indexation_keys: &[I],
message_ids: &[MessageId]
) -> Result<Vec<Message>>

Find all messages by provided message IDs and/or indexation_keys.

pub fn get_balance<'a>(&'a self, seed: &'a Seed) -> GetBalanceBuilder<'a>

Return the balance for a provided seed and its wallet chain account index. Addresses with balance must be consecutive, so this method will return once it encounters a zero balance address.

pub async fn get_address_balances(
&self,
addresses: &[String]
) -> Result<Vec<BalanceAddressResponse>>

Return the balance in iota for the given addresses; No seed or security level needed to do this since we are only checking and already know the addresses.

pub fn bech32_to_hex(bech32: &str) -> Result<String>

Transforms bech32 to hex

pub async fn hex_to_bech32(
&self,
hex: &str,
bech32_hrp: Option<&str>
) -> Result<String>

Transforms hex to bech32

pub fn parse_bech32_address(address: &str) -> Result<Address>

Returns a valid Address parsed from a String.

pub fn is_address_valid(address: &str) -> bool

Checks if a String address is valid.

pub async fn retry(
&self,
message_id: &MessageId
) -> Result<(MessageId, Message)>

Retries (promotes or reattaches) a message for provided message id. Message should only be retried only if they are valid and haven't been confirmed for a while.

pub async fn retry_until_included(
&self,
message_id: &MessageId,
interval: Option<u64>,
max_attempts: Option<u64>
) -> Result<Vec<(MessageId, Message)>>

Retries (promotes or reattaches) a message for provided message id until it's included (referenced by a milestone). Default interval is 5 seconds and max attempts is 10. Returns reattached messages